refactor: session_id
